+++ src/containers/AdminTagEditor/index.jsx
@@ -9,16 +9,20 @@ import TextField from "material-ui/TextField";
 import Toggle from "material-ui/Toggle";
 import FlatButton from "material-ui/FlatButton";
 import ContentAddIcon from "material-ui/svg-icons/content/add";
+import ColorPicker from "material-ui-color-picker";
 
-import { withOperations } from "../hoc/with-operations";
+import { formatErrorMessage, withOperations } from "../hoc/with-operations";
 import LoadingIndicator from "../../components/LoadingIndicator";
 import TagEditorList from "./TagEditorList";
 import theme from "../../styles/theme";
+import ConfirmationStepsEditor from "./ConfirmationStepsEditor";
+import GSScriptField from "../../components/forms/GSScriptField";
 
 class AdminTagEditor extends Component {
   state = {
     editingTag: undefined,
     isWorking: false,
+    isEditingSteps: false,
     error: undefined
   };
 
@@ -36,7 +40,7 @@ class AdminTagEditor extends Component {
         description: "",
         textColor: "",
         backgroundColor: "",
-        confirmationSteps: [[]],
+        confirmationSteps: [],
         onApplyScript: "",
         webhookUrl: "",
         isAssignable: true
@@ -45,7 +49,8 @@ class AdminTagEditor extends Component {
 
   handleEditTag = tagId => this.setState({ editingTag: this.getTag(tagId) });
 
-  handleCancelEditTag = () => this.setState({ editingTag: undefined });
+  handleCancelEditTag = () =>
+    this.setState({ editingTag: undefined, isEditingScript: false });
 
   handleSaveTag = async () => {
     const { editingTag } = this.state;
@@ -65,7 +70,7 @@ class AdminTagEditor extends Component {
       const result = await this.props.mutations.saveTag(tag);
       if (result.errors) throw new Error(result.errors);
     } catch (error) {
-      this.setState({ error: error.message });
+      this.setState({ error: formatErrorMessage(error.message) });
     } finally {
       this.setState({ isWorking: false });
       this.handleCancelEditTag();
@@ -78,7 +83,7 @@ class AdminTagEditor extends Component {
       const result = await this.props.mutations.deleteTag(tagId);
       if (result.errors) throw new Error(result.errors);
     } catch (error) {
-      this.setState({ error: error.message });
+      this.setState({ error: formatErrorMessage(error.message) });
     } finally {
       this.setState({ isWorking: false });
     }
@@ -90,12 +95,70 @@ class AdminTagEditor extends Component {
     this.setState({ editingTag });
   };
 
+  handleOpenScriptEditor = () => {
+    this.setState({ isEditingScript: !this.state.isEditingScript });
+  };
+
+  handleEditTagScript = script => {
+    this.setState({
+      editingTag: { ...this.state.editingTag, onApplyScript: script },
+      isEditingScript: false
+    });
+  };
+
+  handleEditTextColor = color => {
+    this.setState({
+      editingTag: { ...this.state.editingTag, textColor: color }
+    });
+  };
+
+  handleEditBackgroundColor = color => {
+    this.setState({
+      editingTag: { ...this.state.editingTag, backgroundColor: color }
+    });
+  };
+
+  handleEditWebhookUrl = url => {
+    this.setState({
+      editingTag: { ...this.state.editingTag, webhookUrl: url }
+    });
+  };
+
+  handleToggleStepsEditorOpen = () => {
+    this.setState({ isEditingSteps: !this.state.isEditingSteps });
+  };
+
+  handleSaveConfirmationStep = newStep => {
+    const { editingTag } = this.state;
+    const newConfirmationSteps = [...editingTag.confirmationSteps, newStep];
+    this.setState({
+      editingTag: {
+        ...this.state.editingTag,
+        confirmationSteps: newConfirmationSteps
+      }
+    });
+  };
+
+  handleDeleteConfirmationStep = stepIndex => {
+    const { editingTag } = this.state;
+    let newConfirmationSteps = [...editingTag.confirmationSteps];
+    newConfirmationSteps.splice(stepIndex, 1);
+    this.setState({
+      editingTag: {
+        ...this.state.editingTag,
+        confirmationSteps: newConfirmationSteps
+      }
+    });
+  };
+
   render() {
     const { organizationTags } = this.props;
-    const { editingTag, isWorking, error } = this.state;
+    const { editingTag, isWorking, error, isEditingSteps } = this.state;
 
     if (organizationTags.loading) return <LoadingIndicator />;
-    if (organizationTags.errors) return <p>{organizationTags.errors}</p>;
+    if (organizationTags.errors) {
+      return <PrettyErrors errors={organizationTags.errors} />;
+    }
 
     const { tagList } = organizationTags.organization;
 
@@ -105,6 +168,10 @@ class AdminTagEditor extends Component {
       <FlatButton label="Cancel" onClick={this.handleCancelEditTag} />,
       <FlatButton label={tagVerb} primary={true} onClick={this.handleSaveTag} />
     ];
+
+    // Custom fields are campaign-specific and thus cannot be used in Tag scripts.
+    // However, this is a required prop for GSScriptField
+    const customFields = [""];
 
     const errorActions = [
       <FlatButton label="Ok" primary={true} onClick={this.handleCancelError} />
@@ -125,44 +192,95 @@ class AdminTagEditor extends Component {
           <ContentAddIcon />
         </FloatingActionButton>
         {editingTag && (
-          <Dialog
-            title={`${tagVerb} Tag`}
-            actions={actions}
-            modal={false}
-            open={true}
-            onRequestClose={this.handleCancelEditTag}
-          >
-            <TextField
-              name="title"
-              floatingLabelText="Tag title"
-              value={editingTag.title || ""}
-              onChange={this.createTagEditorHandle}
+          <div>
+            <Dialog
+              title={`${tagVerb} Tag`}
+              actions={actions}
+              modal={false}
+              open={true}
+              onRequestClose={this.handleCancelEditTag}
+            >
+              <div
+                style={{
+                  display: "flex",
+                  flexDirection: "row",
+                  justifyContent: "space-between"
+                }}
+              >
+                <div>
+                  <TextField
+                    name="title"
+                    floatingLabelText="Tag title"
+                    value={editingTag.title || ""}
+                    onChange={this.createTagEditorHandle}
+                  />
+                  <GSScriptField
+                    name="Script"
+                    label="Tag script"
+                    context="tagEditor"
+                    customFields={customFields}
+                    value={editingTag.onApplyScript || ""}
+                    onChange={this.handleEditTagScript}
+                    onClick={this.handleOpenScriptEditor}
+                  />
+                  <br />
+                  <Toggle
+                    name="isAssignable"
+                    label="Allow assignment?"
+                    toggled={editingTag.isAssignable}
+                    onToggle={this.createTagEditorHandle}
+                  />
+                </div>
+                <div>
+                  <TextField
+                    name="description"
+                    floatingLabelText="Tag description"
+                    multiLine={true}
+                    value={editingTag.description || ""}
+                    onChange={this.createTagEditorHandle}
+                  />
+                  <ColorPicker
+                    name="Text Color"
+                    floatingLabelText="Text color"
+                    defaultValue={editingTag.textColor}
+                    value={editingTag.textColor || ""}
+                    onChange={this.handleEditTextColor}
+                  />
+                  <ColorPicker
+                    name="Background Color"
+                    floatingLabelText="Background color"
+                    defaultValue={editingTag.backgroundColor}
+                    value={editingTag.backgroundColor || ""}
+                    onChange={this.handleEditBackgroundColor}
+                  />
+                </div>
+              </div>
+              <div style={{ display: 'flex', alignItems: 'center'}}>
+                <span>Tag confirmation steps: {editingTag.confirmationSteps.length}</span>
+                <FlatButton
+                  label="Manage steps"
+                  onClick={this.handleToggleStepsEditorOpen}
+                  primary
+                  style={{ marginLeft: 8 }}
+                />
+              </div>
+              <TextField
+                name="webhookUrl"
+                floatingLabelText="Webhook url"
+                hintText="If set, a request will be sent to this URL whenever this tag is applied."
+                value={editingTag.webhookUrl || ""}
+                onChange={this.createTagEditorHandle}
+                fullWidth
+              />
+            </Dialog>
+            <ConfirmationStepsEditor
+              confirmationSteps={editingTag.confirmationSteps}
+              handleSaveStep={this.handleSaveConfirmationStep}
+              handleDeleteStep={this.handleDeleteConfirmationStep}
+              handleToggleStepsEditorOpen={this.handleToggleStepsEditorOpen}
+              open={isEditingSteps}
             />
-            <br />
-            <TextField
-              name="description"
-              floatingLabelText="Tag description"
-              multiLine={true}
-              value={editingTag.description || ""}
-              onChange={this.createTagEditorHandle}
-            />
-            <br />
-            <TextField
-              name="onApplyScript"
-              floatingLabelText="Script"
-              multiLine={true}
-              value={editingTag.onApplyScript || ""}
-              onChange={this.createTagEditorHandle}
-            />
-            <br />
-            <br />
-            <Toggle
-              name="isAssignable"
-              label="Allow assignment?"
-              toggled={editingTag.isAssignable}
-              onToggle={this.createTagEditorHandle}
-            />
-          </Dialog>
+          </div>
         )}
         <Dialog
           title="Error"
@@ -194,8 +312,12 @@ const queries = {
             title
             description
             isSystem
-            onApplyScript
             isAssignable
+            onApplyScript
+            textColor
+            backgroundColor
+            webhookUrl
+            confirmationSteps
             createdAt
           }
         }