+++ src/server/auth-passport.js
@@ -1,5 +1,5 @@
 import express from "express";
-import passport from "passport";
+import passport from "@passport-next/passport";
 import Auth0Strategy from "passport-auth0";
 import passportSlack from "@aoberoi/passport-slack";
 import { Strategy as LocalStrategy } from "passport-local";
@@ -16,7 +16,6 @@ const {
   BASE_URL,
   AUTOJOIN_ORG_UUID,
   SLACK_TEAM_NAME,
-  SLACK_TEAM_ID,
   SLACK_CLIENT_ID,
   SLACK_CLIENT_SECRET,
   SLACK_SCOPES,
@@ -95,10 +94,12 @@ function setupSlackPassport() {
   );
 
   const handleLogin = async (req, res) => {
+    logger.warn('98: Inside handle login with', req.user)
     const user = req.user;
     // set slack_id to auth0Id to avoid changing the schema
     const auth0Id = user && user.id;
     if (!auth0Id) {
+      logger.warn('103: null user')
       throw new Error("Null user in login callback");
     }
     let existingUser = await r
@@ -123,6 +124,7 @@ function setupSlackPassport() {
     }
 
     if (!existingUser) {
+      logger.warn('128: no existing user')
       let first_name, last_name;
       const splitName = user.name ? user.name.split(" ") : ["First", "Last"];
       if (user.first_name && user.last_name) {
@@ -160,6 +162,7 @@ function setupSlackPassport() {
       return redirectPostSignIn(req, res, true);
     }
 
+    logger.warn('166: returning')
     return redirectPostSignIn(req, res);
   };
 
@@ -167,8 +170,7 @@ function setupSlackPassport() {
   app.get(
     "/login",
     passport.authenticate("slack", {
-      scope: SLACK_SCOPES.split(","),
-      team: SLACK_TEAM_ID
+      scope: SLACK_SCOPES.split(",")
     })
   );